+++ engine/src/Core/Particle.ts
@@ -121,6 +121,8 @@ export class Particle implements IParticle {
     readonly retina: IParticleRetinaProps;
     readonly #engine;
 
+    readonly #engine;
+
     constructor(
         engine: Engine,
         readonly id: number,