+++ demo/svelte/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.23.0](https://github.com/matteobruni/tsparticles/compare/svelte-demo@1.22.3...svelte-demo@1.23.0) (2022-02-04)
+
+**Note:** Version bump only for package svelte-demo
+
+
+
+
+
 ## [1.22.3](https://github.com/matteobruni/tsparticles/compare/svelte-demo@1.22.2...svelte-demo@1.22.3) (2022-02-02)
 
 **Note:** Version bump only for package svelte-demo